requirements.txt
APScheduler==3.9.1.post1
async-timeout==4.0.2
beautifulsoup4==4.11.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
Flask==2.2.2
Flask-APScheduler==1.12.4
Flask-Cors==3.0.10
Flask-MQTT==1.1.1
flask-redis==0.4.0
fonttools==4.38.0
gevent==22.10.2
greenlet==2.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.2
numba==0.56.4
numpy==1.23.4
outlier-utils==0.0.3
packaging==21.3
paho-mqtt==1.6.1
pandas==1.5.1
patsy==0.5.3
Pillow==9.3.0
pycparser==2.21
PyMySQL==1.0.2
pyod==1.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.3.4
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.44
statsmodels==0.13.5
threadpoolctl==3.1.0
tzdata==2022.6
tzlocal==4.2
validators==0.20.0
Werkzeug==2.2.2
wincertstore==0.2
wrapt==1.14.1
zipp==3.10.0
zope.event==4.5.0
zope.interface==5.5.1
sranodec==0.0.1
cython==0.29.32 # for compile